worker/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt
# Or use e.g. pip install -U Flask if on ifi server (you need to be in user env)

# Flask Framework
Flask #==1.0.2

# Flask Packages
# Flask-Login==0.4.0
# Flask-Migrate==2.0.2
# Flask-Script==2.0.5
# Flask-SQLAlchemy==2.4.0
# Flask-WTF==0.14.2
# Flask-User==1.0.1.5

# Automated tests
# pytest==3.0.5
# pytest-cov==2.4.0

# Other requirements
numpy
shortuuid

# Pandas
pandas # ==1.3.2
ydata-profiling # ==3.0.0

# Flask
click # ==8.0.4
Flask # ==2.0.3
itsdangerous # ==2.1.2
Jinja2 # ==3.1.1
MarkupSafe # ==2.1.1
Werkzeug # ==2.0.3


# Azure
azure-storage-blob
azure-data-tables
azure-storage-queue


# Juice

openai